% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decadata.R
\docType{data}
\name{decadata}
\alias{decadata}
\title{Decadata}
\format{
\subsection{\code{decadata}}{

A data frame with 10 rows and 1 column:
\describe{
\item{a}{integers 1 through 10}
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
decadata
}
\description{
Numbers 1-10
}
\keyword{datasets}
