% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rob.R
\name{print_rob}
\alias{print_rob}
\title{Print Rob}
\source{
The idea/data came from \href{https://www.google.com}{here}
}
\usage{
print_rob(a, b)
}
\arguments{
\item{a}{a string}

\item{b}{a string}
}
\value{
What does it return?
}
\description{
\code{print_rob} prints Rob
}
\details{
Provide more info.

Roxygen workflow:
\itemize{
\item add roxygen block
\item devtools:document()
\item ?print_rob()
\item Repeat
}
}
\examples{
print_rob(a = "Take", b = "home.")
}
\seealso{
Other functions
}
